refactor : path 클래스 작성에 따른 리팩토링
