+++ src/main/java/janggi/domain/mouveRule/AdvisorMoveRule.java
@@ -2,7 +2,8 @@ package janggi.domain.mouveRule;
 
 import janggi.domain.Direction;
 import janggi.domain.board.BoardView;
-import janggi.domain.vo.Position;
+import janggi.domain.vo.position.Path;
+import janggi.domain.vo.position.Position;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,18 +13,6 @@ public class AdvisorMoveRule implements MoveRule {
 
     @Override
     public boolean canMove(Position from, Position to, BoardView board) {
-        return candidatePositions(from).contains(to);
-    }
-
-    private List<Position> candidatePositions(Position from) {
-        List<Position> positions =  new ArrayList<>();
-
-        for (Direction direction : ADVISOR_PATHS) {
-            if (from.hasNext(direction)) {
-                positions.add(from.nextPosition(direction));
-            }
-        }
-
-        return positions;
+        return Path.candidatePositions(from, ADVISOR_PATHS).contains(to);
     }
 }
