+++ src/main/java/janggi/domain/mouveRule/MoveRule.java
@@ -1,7 +1,7 @@
 package janggi.domain.mouveRule;
 
 import janggi.domain.board.BoardView;
-import janggi.domain.vo.Position;
+import janggi.domain.vo.position.Position;
 
 public interface MoveRule {
     boolean canMove(Position from, Position to, BoardView board);
