+++ src/main/java/janggi/domain/mouveRule/SoldierMoveRule.java
@@ -3,7 +3,8 @@ package janggi.domain.mouveRule;
 import janggi.domain.Direction;
 import janggi.domain.board.BoardView;
 import janggi.domain.piece.Team;
-import janggi.domain.vo.Position;
+import janggi.domain.vo.position.Path;
+import janggi.domain.vo.position.Position;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,18 +18,6 @@ public final class SoldierMoveRule implements MoveRule {
 
     @Override
     public boolean canMove(Position from, Position to, BoardView board) {
-        return candidatePositions(from).contains(to);
-    }
-
-    private List<Position> candidatePositions(Position from) {
-        List<Position> positions =  new ArrayList<>();
-
-        for (Direction direction : SOLDIER_PATHS) {
-            if (from.hasNext(direction)) {
-                positions.add(from.nextPosition(direction));
-            }
-        }
-
-        return positions;
+        return Path.candidatePositions(from, SOLDIER_PATHS).contains(to);
     }
 }
